+++ src/lib/terkin/telemetry.py
@@ -390,8 +390,8 @@ class TelemetryTransportHTTP:
         log.info('Sending HTTP request to %s', self.uri)
         log.info('HTTP payload: %s', dataframe.payload_out)
 
-        import uurequests
-        response = uurequests.post(self.uri, data=dataframe.payload_out.encode(), headers={'Content-Type': self.content_type})
+        import urequests
+        response = urequests.post(self.uri, data=dataframe.payload_out.encode(), headers={'Content-Type': self.content_type})
         if response.status_code in [200, 201]:
             return True
         else: